# Reduce calls to date_default_timezone_get()