@greg0ire 2021-05-27
Converting to draft until I find out why `date_default_timezone_get` was so slow on that system.