@stof 2021-05-28
> Converting to draft until I find out why `date_default_timezone_get` was so slow on that system.

Does this system define the timezone in its php.ini or does it rely on guessing it from the system ?