@greg0ire 2021-05-31
@stof yes, the timezone is defined in that file.
![Untitled](https://user-images.githubusercontent.com/657779/120164743-2a908b00-c1fb-11eb-8ba2-7b3fd0696dd6.png)